Factor out FtTestApp.
