+++ searchlib/src/tests/nativerank/nativerank_test.cpp
@@ -10,8 +10,8 @@
 #include <vespa/searchlib/fef/functiontablefactory.h>
 #include <vespa/searchlib/fef/test/plugin/setup.h>
 #include <vespa/vespalib/util/stringfmt.h>
-#include <vespa/searchlib/fef/test/ftlib.h>
 #include <vespa/searchlib/fef/test/dummy_dependency_handler.h>
+#include <vespa/searchlib/test/ft_test_app.h>
 
 #include <vespa/log/log.h>
 LOG_SETUP("nativerank_test");
