+++ searchlib/src/vespa/searchlib/test/CMakeLists.txt
@@ -4,6 +4,8 @@ vespa_add_library(searchlib_test
     attribute_builder.cpp
     document_weight_attribute_helper.cpp
     doc_builder.cpp
+    ft_test_app.cpp
+    ft_test_app_base.cpp
     imported_attribute_fixture.cpp
     initrange.cpp
     make_attribute_map_lookup_node.cpp
