+++ searchlib/src/vespa/searchlib/test/ft_test_app.cpp
@@ -1,0 +1,5 @@
+// Copyright Vespa.ai. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
+
+#include "ft_test_app.h"
+
+FtTestApp::~FtTestApp() = default;
